add url encoding/decoding functions
