+++ functions.go
@@ -226,6 +226,8 @@ var genericMap = map[string]interface{}{
 	"b64dec": base64decode,
 	"b32enc": base32encode,
 	"b32dec": base32decode,
+	"urlEnc": urlEncode,
+	"urlDec": urlDecode,
 
 	// Data Structures:
 	"tuple":  list, // FIXME: with the addition of append/prepend these are no longer immutable.
