+++ strings.go
@@ -1,8 +1,6 @@
 package sprig
 
 import (
-	"encoding/base32"
-	"encoding/base64"
 	"fmt"
 	"reflect"
 	"strconv"
@@ -10,30 +8,6 @@ import (
 
 	util "github.com/aokoli/goutils"
 )
-
-func base64encode(v string) string {
-	return base64.StdEncoding.EncodeToString([]byte(v))
-}
-
-func base64decode(v string) string {
-	data, err := base64.StdEncoding.DecodeString(v)
-	if err != nil {
-		return err.Error()
-	}
-	return string(data)
-}
-
-func base32encode(v string) string {
-	return base32.StdEncoding.EncodeToString([]byte(v))
-}
-
-func base32decode(v string) string {
-	data, err := base32.StdEncoding.DecodeString(v)
-	if err != nil {
-		return err.Error()
-	}
-	return string(data)
-}
 
 func abbrev(width int, s string) string {
 	if width < 4 {
